Support consumer mode for simplified version of UI
